[Refactor/#48] Feat: 게시물 관련 코드 구조 개선 및 메타데이터 처리 방식 수정
