+++ mdx.d.ts
@@ -1,0 +1,5 @@
+import type { PostMeta } from "@/lib/types/post";
+
+declare module "*.mdx" {
+  export const meta: PostMeta;
+}
